pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "apaudel96", email = "apaudel96@ymail.com"},
]
dependencies = [
    "flask>=2.1.3",
    "peewee>=3.15.1",
    "python-dotenv>=0.20.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.6.0",
    "pylama>=8.3.8",
]
